add signal strength icons

+++ CustomTypes.h
@@ -12,7 +12,7 @@ Q_DECLARE_METATYPE(ManagedObject)
 using ManagedObjectList = QMap<QDBusObjectPath,ManagedObject>;
 Q_DECLARE_METATYPE(ManagedObjectList);
 
-using OrderedNetworkList = QList<QPair<QDBusObjectPath,int>>;
+using OrderedNetworkList = QList<QPair<QDBusObjectPath,int16_t>>;
 Q_DECLARE_METATYPE(OrderedNetworkList);
 
 using LevelsList = QList<int16_t>;
